refactor: refactor table definition structure
